+++ core/node/base_token_adjuster/src/lib.rs
@@ -1,8 +1,10 @@
 pub use self::{
-    base_token_ratio_persister::{BaseTokenRatioPersister, BaseTokenRatioPersisterL1Params},
+    base_token_l1_behaviour::{BaseTokenL1Behaviour, UpdateOnL1Params},
+    base_token_ratio_persister::BaseTokenRatioPersister,
     base_token_ratio_provider::{DBBaseTokenRatioProvider, NoOpRatioProvider},
 };
 
+mod base_token_l1_behaviour;
 mod base_token_ratio_persister;
 mod base_token_ratio_provider;
 mod metrics;
