+++ core/node/node_framework/src/implementations/layers/base_token/base_token_ratio_persister.rs
@@ -1,9 +1,11 @@
-use zksync_base_token_adjuster::{BaseTokenRatioPersister, BaseTokenRatioPersisterL1Params};
+use std::sync::{Arc, RwLock};
+
+use zksync_base_token_adjuster::{BaseTokenL1Behaviour, BaseTokenRatioPersister, UpdateOnL1Params};
 use zksync_config::{
     configs::{base_token_adjuster::BaseTokenAdjusterConfig, wallets::Wallets},
     ContractsConfig,
 };
-use zksync_contracts::chain_admin_contract;
+use zksync_contracts::{chain_admin_contract, getters_facet_contract};
 use zksync_eth_client::clients::PKSigningClient;
 use zksync_types::L1ChainId;
 
@@ -83,38 +85,44 @@ impl WiringLayer for BaseTokenRatioPersisterLayer {
             .base_token_addr
             .expect("base token address is not set");
 
-        let l1_params =
-            self.wallets_config
-                .token_multiplier_setter
-                .map(|token_multiplier_setter| {
-                    let tms_private_key = token_multiplier_setter.wallet.private_key();
-                    let tms_address = token_multiplier_setter.wallet.address();
-                    let EthInterfaceResource(query_client) = input.eth_client;
+        let l1_behaviour = self
+            .wallets_config
+            .token_multiplier_setter
+            .map(|token_multiplier_setter| {
+                let tms_private_key = token_multiplier_setter.wallet.private_key();
+                let tms_address = token_multiplier_setter.wallet.address();
+                let EthInterfaceResource(query_client) = input.eth_client;
 
-                    let signing_client = PKSigningClient::new_raw(
-                        tms_private_key.clone(),
-                        self.contracts_config.diamond_proxy_addr,
-                        self.config.default_priority_fee_per_gas,
-                        #[allow(clippy::useless_conversion)]
-                        self.l1_chain_id.into(),
-                        query_client.clone().for_component("base_token_adjuster"),
-                    );
-                    BaseTokenRatioPersisterL1Params {
+                let signing_client = PKSigningClient::new_raw(
+                    tms_private_key.clone(),
+                    self.contracts_config.diamond_proxy_addr,
+                    self.config.default_priority_fee_per_gas,
+                    #[allow(clippy::useless_conversion)]
+                    self.l1_chain_id.into(),
+                    query_client.clone().for_component("base_token_adjuster"),
+                );
+                BaseTokenL1Behaviour::UpdateOnL1 {
+                    params: UpdateOnL1Params {
                         eth_client: Box::new(signing_client),
                         gas_adjuster: input.tx_params.0,
                         token_multiplier_setter_account_address: tms_address,
                         chain_admin_contract: chain_admin_contract(),
+                        getters_facet_contract: getters_facet_contract(),
                         diamond_proxy_contract_address: self.contracts_config.diamond_proxy_addr,
                         chain_admin_contract_address: self.contracts_config.chain_admin_addr,
-                    }
-                });
+                        config: self.config.clone(),
+                        last_persisted_l1_ratio: Arc::new(RwLock::new(None)),
+                    },
+                }
+            })
+            .unwrap_or(BaseTokenL1Behaviour::NoOp);
 
         let persister = BaseTokenRatioPersister::new(
             master_pool,
             self.config,
             base_token_addr,
             price_api_client.0,
-            l1_params,
+            l1_behaviour,
         );
 
         Ok(Output { persister })
